+++ aa_intel_tool/static/aa_intel_tool/javascript/aa-intel-tool-dscan.js
@@ -62,7 +62,7 @@ $(() => {
                         {
                             targets: 0,
                             createdCell: (td) => {
-                                $(td).addClass('text-ellipsis');
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
                             }
                         },
                         {
@@ -166,7 +166,7 @@ $(() => {
                         {
                             targets: 0,
                             createdCell: (td) => {
-                                $(td).addClass('text-ellipsis');
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
                             }
                         },
                         {
@@ -270,7 +270,7 @@ $(() => {
                         {
                             targets: 0,
                             createdCell: (td) => {
-                                $(td).addClass('text-ellipsis');
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
                             }
                         },
                         {
@@ -367,8 +367,14 @@ $(() => {
                     ],
                     columnDefs: [
                         {
+                            targets: 0,
+                            createdCell: (td) => {
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
+                            }
+                        },
+                        {
                             targets: 1,
-                            width: 45,
+                            width: 35,
                             createdCell: (td) => {
                                 $(td).addClass('text-end');
                             }
@@ -438,8 +444,14 @@ $(() => {
                     ],
                     columnDefs: [
                         {
+                            targets: 0,
+                            createdCell: (td) => {
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
+                            }
+                        },
+                        {
                             targets: 1,
-                            width: 45,
+                            width: 35,
                             createdCell: (td) => {
                                 $(td).addClass('text-end');
                             }
@@ -507,8 +519,14 @@ $(() => {
                     ],
                     columnDefs: [
                         {
+                            targets: 0,
+                            createdCell: (td) => {
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
+                            }
+                        },
+                        {
                             targets: 1,
-                            width: 45,
+                            width: 35,
                             createdCell: (td) => {
                                 $(td).addClass('text-end');
                             }
@@ -576,8 +594,14 @@ $(() => {
                     ],
                     columnDefs: [
                         {
+                            targets: 0,
+                            createdCell: (td) => {
+                                $(td).addClass('text-ellipsis fix-eve-image-position');
+                            }
+                        },
+                        {
                             targets: 1,
-                            width: 45,
+                            width: 35,
                             createdCell: (td) => {
                                 $(td).addClass('text-end');
                             }